Fix runtime errors, module names
