+++ src/App.tsx
@@ -1,6 +1,5 @@
 import { useState } from 'react';
-import reactLogo from './assets/react.svg';
-import '../public/output.css';
+import '../src/output.css';
 
 function App() {
 	const [count, setCount] = useState(0);
@@ -10,10 +9,10 @@ function App() {
 		<div className="App" /*class=""*/>
 			<div>
 				<a href="https://vitejs.dev" target="_blank">
-					<img src="/vite.svg" className="logo" alt="Vite logo" />
+					{/* <img src="/vite.svg" className="logo" alt="Vite logo" /> */}
 				</a>
 				<a href="https://reactjs.org" target="_blank">
-					<img src={reactLogo} className="logo react" alt="React logo" />
+					{/* <img src='/assets/react.svg' className="logo react" alt="React logo" /> */}
 				</a>
 			</div>
 			<h1>Vite + React</h1>
@@ -22,7 +21,7 @@ function App() {
 					count is {count}
 				</button>
 				<p>
-					Edit <code>src/App.jsx</code> and save to test HMR
+					Edit <code>src/App.tsx</code> and save to test HMR
 				</p>
 			</div>
 			<p className="read-the-docs">
