--- src/components/DarkMode/DarkBtn.jsx
@@ -1,11 +1,0 @@
-
-export default function DarkBtn () {
-  return (
-    <button
-      id="theme-toggle"
-      type="button"
-      class="p-4 my-2 flex text-blue-700 dark:bg-red-500 hover:scale-110 duration-100 hover:duration-100 rounded-lg text-sm ">
-      <span class="p-1 font-bold fa-large fa-moon">Dark</span>
-    </button>
-  )
-}
